Reutilizar componentes no fue una buena idea
